refactor: rename plugin to kubectl-print-env
