+++ cmd/kubectl-print-env/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"github.com/pedrobarco/kubectl-env/cmd/kubectl-env/env"
+	"github.com/pedrobarco/kubectl-print-env/cmd/kubectl-print-env/printenv"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 )
 
 func main() {
-	c := env.NewCmdEnv()
+	c := printenv.NewCmdEnv()
 	_ = c.Execute()
 }
